structure refactoring
parse command and payloads at a time
because old app searching process was O(N) and I want to make it O(1)
so that parse command from original text first and make a lookup table of apps with it.

parse payloads separately would not be efficient but it is effective
It's obvious to give user input text to the app would be fancier than to give developer tokenized list when write a test code for the app.